backend/requirements.txt
Flask-SQLAlchemy
flask-smorest
flask-migrate
python-dotenv
mysqlclient
flask-cors
passlib
flask-jwt-extended